Move ui codes from 'sample-app' to 'sample-ui'
