+++ sample-app/build.gradle
@@ -30,6 +30,7 @@ android {
 
 dependencies {
     implementation project(":neumorphism")
+    implementation project(":sample-ui")
     implementation Libs.Kotlin.stdlib
     implementation Libs.AndroidX.appcompat
     implementation Libs.AndroidX.constraintlayout
