+++ sample-app/src/main/AndroidManifest.xml
@@ -17,9 +17,6 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".ColorSampleActivity" />
-        <activity android:name=".CardSampleActivity" />
-        <activity android:name=".ListSampleActivity" />
     </application>
 
 </manifest>
