+++ sample-ui/src/main/AndroidManifest.xml
@@ -2,4 +2,9 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="soup.neumorphism.sample.ui">
 
+    <application android:theme="@style/AppTheme">
+        <activity android:name=".ColorSampleActivity" />
+        <activity android:name=".CardSampleActivity" />
+        <activity android:name=".ListSampleActivity" />
+    </application>
 </manifest>
