+++ sample-ui/src/main/res/drawable-night/gradation.xml
