+++ sample-ui/src/main/res/drawable-night/ic_daynight_activity.xml
