+++ sample-ui/src/main/res/drawable/gradation.xml
