+++ sample-ui/src/main/res/drawable/ic_daynight_food.xml
