+++ sample-ui/src/main/res/drawable/ic_filter.xml
