+++ sample-ui/src/main/res/drawable/ic_left.xml
