+++ sample-ui/src/main/res/drawable/ic_right.xml
