+++ sample-ui/src/main/res/layout/activity_sample_card.xml
@@ -10,7 +10,7 @@
         android:layout_width="280dp"
         android:layout_height="wrap_content"
         android:layout_gravity="center"
-        app:layout_constraintEnd_toStartOf="@id/pressed_card"
+        app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@id/textview">
