+++ sample-ui/src/main/res/layout/activity_sample_list.xml
