+++ sample-ui/src/main/res/layout/item_card.xml
