+++ sample-ui/src/main/res/values-night/colors.xml
