+++ sample-ui/src/main/res/values/colors.xml
