+++ sample-ui/src/main/res/values/styles.xml
