+++ sample-ui/src/main/res/values/sys_ui.xml
